+++ app/src/main/java/campus/tech/kakao/Model/ResultSearchKeyword.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.Model
 
 import com.google.gson.annotations.SerializedName
