+++ app/src/main/java/campus/tech/kakao/Model/RetrofitClient.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.Model
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
