+++ app/src/main/java/campus/tech/kakao/Model/SQLiteHelper.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.Model
 
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
