+++ app/src/main/java/campus/tech/kakao/View/PlacesAdapter.kt
@@ -1,10 +1,12 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.View
 
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import campus.tech.kakao.Model.Place
+import campus.tech.kakao.map.R
 
 class PlacesAdapter(private var places: List<Place>, private val onItemClick: (String) -> Unit) : RecyclerView.Adapter<PlacesAdapter.PlaceViewHolder>() {
